# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Grain weighing registrations API
  version: 0.3.0
paths:
  /v1/scale-readings:
    post:
      summary: Authenticate a scale and enqueue a reading in Redis Streams
      parameters:
        - in: header
          name: Authorization
          required: true
          schema: { type: string, example: "Bearer device-key" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event_id, scale_id, plate, weight_grams, measured_at]
              properties:
                event_id: { type: string, format: uuid }
                scale_id: { type: string }
                plate: { type: string, description: Normalized to seven uppercase alphanumeric characters }
                weight_grams: { type: integer, format: int64, minimum: 1 }
                measured_at: { type: string, format: date-time }
      responses:
        '202': { description: Redis confirmed XADD to scale-readings }
        '400': { description: Malformed or oversized JSON body }
        '401': { description: Missing or invalid device key }
        '403': { description: Disabled scale or scale/device-key mismatch }
        '422': { description: Business-invalid reading }
        '503': { description: Redis unavailable }
  /v1/branches:
    get: { summary: List branches, responses: { '200': { description: OK } } }
    post: { summary: Create branch, responses: { '201': { description: Created }, '422': { description: Validation error } } }
  /v1/branches/{id}:
    get: { summary: Get branch, responses: { '200': { description: OK }, '404': { description: Not found } } }
    put: { summary: Update active branch, responses: { '200': { description: OK } } }
...
  /v1/trucks:
    get: { summary: List trucks, responses: { '200': { description: OK } } }
    post: { summary: Create truck with normalized unique plate and integer tare grams, responses: { '201': { description: Created }, '409': { description: Duplicate plate } } }
...
  /v1/transport-transactions:
    get: { summary: List transport transactions, responses: { '200': { description: OK } } }
    post: { summary: Open transport transaction with price and margin snapshots, responses: { '201': { description: Created }, '409': { description: Existing OPEN transaction } } }
...
  /v1/transport-transactions/{id}/status:
    patch: { summary: Cancel an OPEN transaction; WEIGHED is reserved for finalization, responses: { '200': { description: Cancelled }, '409': { description: Invalid transition } } }